+++ beetsplug/embyupdate.py
@@ -3,25 +3,38 @@
 """Updates the Emby Library whenever the beets library is changed.
 
     emby:
-        host: localhost
+        host: http://localhost
         port: 8096
         username: user
         password: password
 """
 from __future__ import division, absolute_import, print_function
 
-from beets import config
-from beets.plugins import BeetsPlugin
-from six.moves.urllib.parse import urlencode
-from six.moves.urllib.parse import urljoin, parse_qs, urlsplit, urlunsplit
 import hashlib
 import requests
+
+from six.moves.urllib.parse import urlencode
+from six.moves.urllib.parse import urljoin, parse_qs, urlsplit, urlunsplit
+
+from beets import config
+from beets.plugins import BeetsPlugin
 
 
 def api_url(host, port, endpoint):
     """Returns a joined url.
+
+    Takes host, port and endpoint and generates a valid emby API url.
+
+    :param host: Hostname of the emby server
+    :param port: Portnumber of the emby server
+    :param endpoint: API endpoint
+    :type host: str
+    :type port: int
+    :type endpoint: str
+    :returns: Full API url
+    :rtype: str
     """
-    joined = urljoin('http://{0}:{1}'.format(host, port), endpoint)
+    joined = urljoin('{0}:{1}'.format(host, port), endpoint)
     scheme, netloc, path, query_string, fragment = urlsplit(joined)
     query_params = parse_qs(query_string)
 
@@ -33,6 +46,13 @@ def api_url(host, port, endpoint):
 
 def password_data(username, password):
     """Returns a dict with username and its encoded password.
+
+    :param username: Emby username
+    :param password: Emby password
+    :type username: str
+    :type password: str
+    :returns: Dictionary with username and encoded password
+    :rtype: dict
     """
     return {
         'username': username,
@@ -43,24 +63,45 @@ def password_data(username, password):
 
 def create_headers(user_id, token=None):
     """Return header dict that is needed to talk to the Emby API.
+
+    :param user_id: Emby user ID
+    :param token: Authentication token for Emby
+    :type user_id: str
+    :type token: str
+    :returns: Headers for requests
+    :rtype: dict
     """
-    headers = {
-        'Authorization': 'MediaBrowser',
-        'UserId': user_id,
-        'Client': 'other',
-        'Device': 'empy',
-        'DeviceId': 'beets',
-        'Version': '0.0.0'
-    }
+    headers = {}
+
+    authorization = (
+        'MediaBrowser UserId="{user_id}", '
+        'Client="other", '
+        'Device="beets", '
+        'DeviceId="beets", '
+        'Version="0.0.0"'
+    ).format(user_id=user_id)
+
+    headers['x-emby-authorization'] = authorization
 
     if token:
-        headers['X-MediaBrowser-Token'] = token
+        headers['x-mediabrowser-token'] = token
 
     return headers
 
 
 def get_token(host, port, headers, auth_data):
     """Return token for a user.
+
+    :param host: Emby host
+    :param port: Emby port
+    :param headers: Headers for requests
+    :param auth_data: Username and encoded password for authentication
+    :type host: str
+    :type port: int
+    :type headers: dict
+    :type auth_data: dict
+    :returns: Access Token
+    :rtype: str
     """
     url = api_url(host, port, '/Users/AuthenticateByName')
     r = requests.post(url, headers=headers, data=auth_data)
@@ -70,6 +111,15 @@ def get_token(host, port, headers, auth_data):
 
 def get_user(host, port, username):
     """Return user dict from server or None if there is no user.
+
+    :param host: Emby host
+    :param port: Emby port
+    :username: Username
+    :type host: str
+    :type port: int
+    :type username: str
+    :returns: Matched Users
+    :rtype: list
     """
     url = api_url(host, port, '/Users/Public')
     r = requests.get(url)
@@ -84,7 +134,7 @@ class EmbyUpdate(BeetsPlugin):
 
         # Adding defaults.
         config['emby'].add({
-            u'host': u'localhost',
+            u'host': u'http://localhost',
             u'port': 8096
         })